safari-600.7-branch/LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.5_Tokens/7.5.2_Keywords/S7.5.2_A1.10-expected.txt

CONSOLE MESSAGE: line 76: SyntaxError: Unexpected token '='. Expected '(' to start a for-loop header.
S7.5.2_A1.10

PASS Expected parsing failure

TEST COMPLETE
